requirements.txt
click>=8.1.0
openai>=1.40.0
python-dotenv>=1.0.0
